[gh315] Fix credo helper
